Reorganize classes into best fit packages
* Move Matching.kt into the util package
* Move fuzzy matching tests into the util package
