+++ src/main/java/com/google/idea/perf/cachedvaluetracer/CachedValueTracerCommandPredictor.kt
@@ -17,7 +17,7 @@
 package com.google.idea.perf.cachedvaluetracer
 
 import com.google.idea.perf.CommandPredictor
-import com.google.idea.perf.fuzzySearch
+import com.google.idea.perf.util.fuzzySearch
 import com.intellij.openapi.progress.ProgressManager
 
 class CachedValueTracerCommandPredictor: CommandPredictor {
