+++ src/main/java/com/google/idea/perf/methodtracer/MethodTracerCommandPredictor.kt
@@ -17,7 +17,7 @@
 package com.google.idea.perf.methodtracer
 
 import com.google.idea.perf.CommandPredictor
-import com.google.idea.perf.fuzzySearch
+import com.google.idea.perf.util.fuzzySearch
 import com.intellij.openapi.progress.ProgressManager
 
 class MethodTracerCommandPredictor: CommandPredictor {
